feat: move utils common to v1 - v2 - v3
